Add files via upload
In this code, I used my skills to write a simple version of an alarm clock. In this code I used TypeScript for user side security, generator functions, local storage, etc.
It is possible to see the current hour, create an alarm, delete it and turn it on/off.
The list of alarm clocks is sorted in ascending order. That is, 00:00 to 23:59 hours.
I tried to check everything that the user could not create an alarm clock, for example: 56:99 or 87:67...
If the hour or minute digits are missing, the user sees an alert (Please set alarm completely!).
The user can only create unique alarms.
That's all, I hope you all like my simple version of an alarm clock!